[NSE-927] Implement AVX512 optimization selection in Runtime and merge two C2R code files into one. (#937)
* extract and inline functions

* Remove debug info

* Implement AVX512 optimization selection in Runtime.

* Fix Clang-format

* Use aligned AVX

* Add issue comment
